refactor(projects): refactored projects files
refactored projects files and folders along with placing variants into a new file
