+++ .firebase/hosting.YnVpbGQ.cache
@@ -1,27 +1,28 @@
-index.html,1652398557861,ee7c3afbf9da348c510a20a069cd72161e40d5e20257bf63982631f620e63cb6
-asset-manifest.json,1652398557867,cc683d1aa19a0720754f9c35ce92633e3e451bc65896c33dd420c461cabd4850
-manifest.json,1652398548240,341d52628782f8ac9290bbfc43298afccb47b7cbfcee146ae30cf0f46bc30900
-404.html,1652398721125,05cbc6f94d7a69ce2e29646eab13be2c884e61ba93e3094df5028866876d18b3
-favicon.ico,1652398548239,b72f7455f00e4e58792d2bca892abb068e2213838c0316d6b7a0d6d16acd1955
-robots.txt,1652398548240,391d14b3c2f8c9143a27a28c7399585142228d4d1bdbe2c87ac946de411fa9a2
-static/css/main.d4f3ba59.css,1652398557862,cc64a3d6ea4ab3f3e5a40774bd9b8b5a890f626ad1ed0cef6eee3a78e22978c7
-logo192.png,1652398548239,caff018b7f1e8fd481eb1c50d75b0ef236bcd5078b1d15c8bb348453fee30293
-static/js/787.1ed0c42c.chunk.js,1652398557862,a0032a581a6f10e5be393fc5bf09a7c9ad026b1d03e05ea5fd446b8b342d42c1
-logo512.png,1652398548240,191fc21360b4ccfb1cda11a1efb97f489ed22672ca83f4064316802bbfdd750e
-static/js/main.8503771f.js.LICENSE.txt,1652398557862,bfbf3a4b9e414262e302b9c850733cb7cfafa9577cb62b814d04300a511215fc
-static/css/main.d4f3ba59.css.map,1652398557867,f2e2cd4073ef34f8ff974c5eff93acb374ecfe5f1d3a6d4b055f1dc71e40c630
-static/js/787.1ed0c42c.chunk.js.map,1652398557868,5b85ae311399f67f2d1868a7d371a3db5ce90e2d43e4c3ab85b47ce2384a1b46
-static/media/arrowIcon.ab24f3efa2f74e0c2d557c325af9d0ec.svg,1652398557862,7e987ca50e11d7b0a028e27466026518a7e4b0f2b7944edd2655535c8371eac6
-static/js/937.cc77dee8.chunk.js,1652398557862,37b4e282b7cafde0925fc7734cc1bbfecfd8fc8f8852baf70e3a529fe02d4cce
-static/media/deskIcon.99c7bef9d23d844d78e7d7bdaa8492cf.svg,1652398557862,add85e8a56671817f78b536524b24a2fecd5aa9f8a1b7cc313d89973d99a1ddc
-static/media/aboutIcon.3ae31027d1bec7d0053393a50357cf2d.svg,1652398557861,db534fc01bab7f0f952fdad7bb137a5647b1ad28db1fe898519f26c239fadeec
-static/media/githubIcon.60c6f2aec6d5c68cb00e317139e8af6c.svg,1652398557862,c958166a6a12572bd09a0541b0406c66d830a42d1f1832582f2e05631923f95b
-static/media/heroIcon.b6551a9d56b4bf6338504b0fae566c49.svg,1652398557861,e3bca2c0e71a763f87603b7a182da36e99d3bf4287c2b390299da32da2c26d40
-static/media/keyboardIcon.bebb6a0accc78e77c9034dc3e71f07db.svg,1652398557862,77e5ddbcdefd6e068edceef0f7c1f724c6586e20d081abf88b70d419fdfaf44c
-static/media/twitterIcon.17e04658bf5f0ae7ebfa046e1add77d7.svg,1652398557862,51fa0d9762640815409644ab68c4edbad6a6028250b0086f5ccf9ff6afcd6472
-static/js/937.cc77dee8.chunk.js.map,1652398557867,d2ddda978f9ec388442621f1c4fe43263eed28810371a6f46768118afd05315a
-static/js/main.8503771f.js,1652398557862,09d2d5b245074fd2db185725c7acf55107ecf7542caf41a493c96eff0a3cb544
-static/media/gameApp.f39dad814a3826a9d5af.png,1652398557862,45b3db8f10bfaba2b0dec5b2f4c34239b09f2ff762fa40e0cbb37c2a30de5275
-static/media/newsAppImg.a3ee14c3666b7503a51b.png,1652398557862,5b9e8654ff3977b7b9a56d8083771a65b42402b34a27e74e73de3899865c4fc1
-static/media/pokeImg.48230b1672d290bd76fc.png,1652398557862,9ccf3431f0bfb434283351c3e07d9d6ea12ebc00eb00fbb7ae24168f3c94012f
-static/js/main.8503771f.js.map,1652398557868,877bc352214bb2c6ead2a47e8c165c4894bfe392b2dc6899509e831abab3d0d9
+asset-manifest.json,1652475451236,e7d02e6a5bbc82df0e20e40fa6cad9f205cd80ba1ea3b1e1c05ec36d1da71ca3
+index.html,1652475451229,e1440cf26a4ec41816a47adf2970dc8682b3a6c4776f1f84c9fe4016acff326f
+manifest.json,1652475445120,50f4e83481658d16b0dce6859ad0b2acefa5f06da60b141c6e5d5475a7f38777
+robots.txt,1652475445120,391d14b3c2f8c9143a27a28c7399585142228d4d1bdbe2c87ac946de411fa9a2
+static/css/main.dfcb1a1f.css,1652475451231,fce7222bdab4a47c42f570d986cef8e875470222ca35b91f82fdf8d7b8fe017f
+logo192.png,1652475445120,c6461015e9f81b044037750b56c39af7923241e0c5f9edffdd4ebd0466377460
+static/js/787.1ed0c42c.chunk.js,1652475451236,a0032a581a6f10e5be393fc5bf09a7c9ad026b1d03e05ea5fd446b8b342d42c1
+static/css/main.dfcb1a1f.css.map,1652475451237,8289dd7edf827cefd493d053edaef456658e01ff166b133783eb3d4bea47c6fb
+static/js/main.ba644664.js.LICENSE.txt,1652475451231,bfbf3a4b9e414262e302b9c850733cb7cfafa9577cb62b814d04300a511215fc
+static/js/787.1ed0c42c.chunk.js.map,1652475451237,5b85ae311399f67f2d1868a7d371a3db5ce90e2d43e4c3ab85b47ce2384a1b46
+static/js/937.cc77dee8.chunk.js,1652475451237,37b4e282b7cafde0925fc7734cc1bbfecfd8fc8f8852baf70e3a529fe02d4cce
+static/media/aboutIcon.3ae31027d1bec7d0053393a50357cf2d.svg,1652475451231,db534fc01bab7f0f952fdad7bb137a5647b1ad28db1fe898519f26c239fadeec
+static/media/arrowIcon.ab24f3efa2f74e0c2d557c325af9d0ec.svg,1652475451231,7e987ca50e11d7b0a028e27466026518a7e4b0f2b7944edd2655535c8371eac6
+favicon.ico,1652475445120,8720fb8a5cdaca042fdf5e8ead0524e0c22067620607b9d40e87ee29a25795e3
+static/media/keyboardIcon.bebb6a0accc78e77c9034dc3e71f07db.svg,1652475451231,77e5ddbcdefd6e068edceef0f7c1f724c6586e20d081abf88b70d419fdfaf44c
+static/media/deskIcon.99c7bef9d23d844d78e7d7bdaa8492cf.svg,1652475451231,add85e8a56671817f78b536524b24a2fecd5aa9f8a1b7cc313d89973d99a1ddc
+static/media/githubIcon.60c6f2aec6d5c68cb00e317139e8af6c.svg,1652475451230,c958166a6a12572bd09a0541b0406c66d830a42d1f1832582f2e05631923f95b
+static/media/heroIcon.b6551a9d56b4bf6338504b0fae566c49.svg,1652475451231,e3bca2c0e71a763f87603b7a182da36e99d3bf4287c2b390299da32da2c26d40
+static/media/phoneIcon.a4ee949d5e1396b96f59449a2762702a.svg,1652475451231,04e48291666f10ec18ca608e7ab3f7c989e1c1e0dc88ca7c1bd1fd79d1acc6d6
+static/media/twitterIcon.17e04658bf5f0ae7ebfa046e1add77d7.svg,1652475451231,51fa0d9762640815409644ab68c4edbad6a6028250b0086f5ccf9ff6afcd6472
+logo512.png,1652475445120,99cfb579a01f7e9b688228c3a6616ecf454a4b75a016c84e184962833bcc64c4
+static/js/937.cc77dee8.chunk.js.map,1652475451237,d2ddda978f9ec388442621f1c4fe43263eed28810371a6f46768118afd05315a
+static/js/main.ba644664.js,1652475451231,0022dc8ed2a161a6b19f6e3e19c8b58cd024348e782289f8ae55b4d604c65c8c
+static/media/gameApp.f39dad814a3826a9d5af.png,1652475451231,45b3db8f10bfaba2b0dec5b2f4c34239b09f2ff762fa40e0cbb37c2a30de5275
+static/media/newsAppImg.a3ee14c3666b7503a51b.png,1652475451231,5b9e8654ff3977b7b9a56d8083771a65b42402b34a27e74e73de3899865c4fc1
+static/media/pokeImg.48230b1672d290bd76fc.png,1652475451231,9ccf3431f0bfb434283351c3e07d9d6ea12ebc00eb00fbb7ae24168f3c94012f
+static/media/amazon.07330f104bb7501b2fd0.png,1652475451230,9ad70614a450674ea74e0f28ba1220d46b05ae7ee6f600d4080bb0a3086a14cc
+static/js/main.ba644664.js.map,1652475451237,746012791780459ea00c3fa390d0a25eb491aff53a7ed7dc3dc4c7b973680d4b
