+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "portfolio-e5bdc"
+  }
+}
