+++ src/components/Projects/Variants.js
@@ -1,0 +1,19 @@
+export const projectsVariant = {
+  hidden: {
+    y: "20vh",
+    opacity: 0,
+  },
+  visible: {
+    y: 0,
+    opacity: 1,
+    transition: {
+      duration: 0.7,
+    },
+  },
+};
+
+export const singleProjVariant = {
+  hover: {
+    scale: 1.05,
+  },
+};
